+++ src/Components/Layout/Input/GPSInput/__LocalComponents__/__CheckboxOption__.tsx
@@ -6,7 +6,40 @@ import ConfigService from '@Services/ConfigService';
 import Checkbox from '@Components/Layout/Button/Checkbox';
 import H3 from '@Components/Layout/Text/H3';
 
-export default function __CheckboxOption__(props: {
+export default function CheckboxOptions(props: {
+  gpsON: boolean
+  editMode: boolean
+  enableCoordinate: boolean
+  enableAltitude: boolean
+  onToogle_Coordinate: (checked: boolean) => void
+  onToogle_Altitude: (checked: boolean) => void
+}) {
+
+  if (props.editMode === false) {
+    return <></>;
+  }
+
+  return (
+    <View
+      style={{
+        gap: 5,
+      }}
+    >
+      <Option
+        title="Coordinate"
+        value={props.enableCoordinate}
+        onChange={(checked) => props.onToogle_Coordinate(checked)}
+      />
+      <Option
+        title="Altitude"
+        value={props.enableAltitude}
+        onChange={(checked) => props.onToogle_Altitude(checked)}
+      />
+    </View>
+  );
+}
+
+function Option(props: {
   title: string
   value: boolean
   onChange: (checked: boolean) => void