+++ src/Components/Layout/Input/GPSInput/__LocalComponents__/index.tsx
@@ -1,12 +1,10 @@
-import __CheckboxOption__ from './__CheckboxOption__';
-import __Root__ from './__Root__';
+import CheckboxOptions from './__CheckboxOption__';
 import __LoadingFeedback_ from './__LoadingFeedback__';
 import __Display_Data__ from './__Display_Data__';
 import __Display_RealTimeAccuracy__ from './__Display_RealTimeAccuracy__';
 
-export const LocalComponent = {
-  Root: __Root__,
-  CheckboxOption: __CheckboxOption__,
+export const LC = {
+  CheckboxOption: CheckboxOptions,
   DisplayData: __Display_Data__,
   Display_RealTimeAccuracy: __Display_RealTimeAccuracy__,
   Loading: __LoadingFeedback_,