+++ docs/ReleaseNotes.md
@@ -1,4 +1,10 @@
-# 0.1.0.0
+# 0.1.1.7 GPSInput Edit iconButton
 
-- Basic Layout
-- local storage implemented
+- Removed options when GPSWatcherService is instantiated or when a new gpsData is setted inside it.
+- iconButton abstracted into its own component.
+- added edit icon button to GPSInput.
+- changed the render conditions for iconButtons on GPSInput
+- New Input Root Component Created for all Inputs to use.
+- Now local components namespaces have the alias "LC".
+- Checkbox Option abstracted into its own component.
+- GPSInput data and accuracy display design refactor.